# docs/openapi/v1.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/access/passes": {
      "get": {
        "operationId": "AccessController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Access"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "AccessController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePassDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Access"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/access/passes/{id}/revoke": {
      "post": {
        "operationId": "AccessController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Access"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/access/passes/{id}/qr": {
      "get": {
        "operationId": "AccessController_qr",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Access"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/access/validate": {
      "post": {
        "operationId": "AccessController_validate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateAccessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessEvent"
                }
              }
            }
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Access"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "SIGRA API",
    "description": "Official SIGRA Phase 0 HTTP contract",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreatePassDto": {
        "type": "object",
        "properties": {
          "validDays": {
            "type": "number",
            "minimum": 1,
            "maximum": 30
          }
        },
        "required": [
          "validDays"
        ]
      },
      "ValidateAccessDto": {
        "type": "object",
        "properties": {
          "qrPayload": {
            "type": "string"
          },
          "clientEventId": {
            "type": "string",
            "format": "uuid"
          },
          "direction": {
            "type": "string",
            "enum": [
              "ENTRY",
              "EXIT"
            ]
          }
        },
        "required": [
          "qrPayload",
          "clientEventId",
          "direction"
        ]
      },
      "AccessEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "clientEventId": {
            "type": "string"
          },
          "passId": {
            "type": "string",
            "nullable": true
          },
          "guardId": {
            "type": "string"
          },
          "residentId": {
            "type": "string",
            "nullable": true
          },
          "unitId": {
            "type": "string",
            "nullable": true
          },
          "decision": {
            "type": "string",
            "enum": [
              "ALLOWED",
              "DENIED"
            ]
          },
          "direction": {
            "type": "string",
            "enum": [
              "ENTRY",
              "EXIT"
            ]
          },
          "reason": {
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "clientEventId",
          "passId",
          "guardId",
          "residentId",
          "unitId",
          "decision",
          "direction",
          "reason",
          "occurredAt",
          "createdAt"
        ]
      },
...
      "Phase0Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "BAD_REQUEST",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "RATE_LIMITED",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "requestId": {
            "type": "string"
          }
        }
      }
    }
  }
}